Move Repository interface to api
