+++ src/main/java/dev/mccue/resolve/maven/MavenRepository.java
@@ -1,5 +1,6 @@
 package dev.mccue.resolve.maven;
 
+import dev.mccue.resolve.api.Repository;
 import dev.mccue.resolve.cache.OSLabel;
 import dev.mccue.resolve.core.*;
